+++ opentelemetry-datadog/CHANGELOG.md
@@ -12,6 +12,12 @@
 
 - Bump opentelemetry version to 0.22, opentelemetry_sdk version to 0.22
 
+### Changed
+
+- allow send all traces to `datadog-agent` with `agent-sampling` feature.
+- allow `datadog-agent` generate metrics from spans for [APM](https://docs.datadoghq.com/tracing/metrics/).
+
+
 ## v0.9.0
 
 ### Changed